requirements.txt
langchain==0.3.17
langchain-community==0.3.16
langchain-huggingface==0.1.2
openai==1.60.2
streamlit==1.25.0
huggingface-hub==0.28.0
sentence-transformers==3.4.1
tokenizers==0.21.0
transformers==4.48.1
torch==2.6.0
scikit-learn==1.6.1
scipy==1.15.1
Pillow==11.1.0
regex==2024.11.6
safetensors==0.5.2
filelock==3.17.0
fsspec==2024.12.0
jinja2==3.1.5
networkx==3.4.2
MarkupSafe==3.0.2
joblib==1.4.2
threadpoolctl==3.5.0
mpmath==1.3.0
sympy==1.13.1
anyio==4.8.0
aiohttp==3.11.11
python-dotenv==1.0.0
